Make int/uint mess sane

+++ buf.go
@@ -12,22 +12,22 @@ type Buffer []rune
 
 func NewBuffer() Buffer { return []rune{} }
 
-func (b *Buffer) Insert(q0 uint, r []rune) {
-	if q0 > uint(len(*b)) {
+func (b *Buffer) Insert(q0 int, r []rune) {
+	if q0 > (len(*b)) {
 		panic("internal error: buffer.Insert: Out of range insertion")
 	}
 	(*b) = append((*b)[:q0], append(r, (*b)[q0:]...)...)
 }
 
-func (b *Buffer) Delete(q0, q1 uint) {
-	if q0 > uint(len(*b)) || q1 > uint(len(*b)) {
+func (b *Buffer) Delete(q0, q1 int) {
+	if q0 > (len(*b)) || q1 > (len(*b)) {
 		panic("internal error: buffer.Delete: Out-of-range Delete")
 	}
 	copy((*b)[q0:], (*b)[q1:])
-	(*b) = (*b)[:uint(len(*b))-(q1-q0)] // Reslice to length
+	(*b) = (*b)[:(len(*b))-(q1-q0)] // Reslice to length
 }
 
-func (b *Buffer) Load(q0 uint, fd *os.File) (n uint, h [sha1.Size]byte, hasNulls bool, err error) {
+func (b *Buffer) Load(q0 int, fd *os.File) (n int, h [sha1.Size]byte, hasNulls bool, err error) {
 	// TODO(flux): Innefficient to load the file, then copy into the slice,
 	// but I need the UTF-8 interpretation.  I could fix this by using a
 	// UTF-8 -> []rune reader on top of the os.File instead.
@@ -41,12 +41,12 @@ func (b *Buffer) Load(q0 uint, fd *os.File) (n uint, h [sha1.Size]byte, hasNulls
 	hasNulls = len(s) != len(d)
 	runes := []rune(s)
 	(*b).Insert(q0, runes)
-	return uint(len(runes)), sha1.Sum(d), hasNulls, err
+	return (len(runes)), sha1.Sum(d), hasNulls, err
 }
 
-func (b *Buffer) Read(q0, n uint) (r []rune) {
+func (b *Buffer) Read(q0, n int) (r []rune) {
 	// TODO(flux): Can I just reslice here, or do I need to copy?
-	if !(q0 <= uint(len(*b)) && q0+n <= uint(len(*b))) {
+	if !(q0 <= (len(*b)) && q0+n <= (len(*b))) {
 		panic("internal error: Buffer.Read")
 	}
 	r = make([]rune, n)
@@ -63,8 +63,8 @@ func (b *Buffer) Reset() {
 	(*b) = (*b)[0:0]
 }
 
-func (b *Buffer) nc() uint {
-	return uint(len(*b))
+func (b *Buffer) nc() int {
+	return len(*b)
 }
 
 func fbufalloc() []rune {
